+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.1.0] - 2026-08-13
 ### Added
 - Animation ResetSpeed function
 - AnimationSequence PlayAsync function
@@ -43,7 +45,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gitignore
 - Initial commit
 
-[Unreleased]: https://github.com/1mbitshorde/AnimationSystem/compare/1.0.0...main
+[Unreleased]: https://github.com/1mbitshorde/AnimationSystem/compare/1.1.0...main
+[1.1.0]: https://github.com/1mbitshorde/AnimationSystem/tree/1.1.0/
 [1.0.0]: https://github.com/1mbitshorde/AnimationSystem/tree/1.0.0/
 [0.3.1]: https://github.com/1mbitshorde/AnimationSystem/tree/0.3.1/
 [0.3.0]: https://github.com/1mbitshorde/AnimationSystem/tree/0.3.0/